+++ src/Dfe.SignIn.Core.Contracts/Users/CancelPendingChangeEmailAddress.cs
@@ -1,4 +1,5 @@
 using Dfe.SignIn.Base.Framework;
+using Dfe.SignIn.Core.Contracts.Features.Users.Exceptions;
 
 namespace Dfe.SignIn.Core.Contracts.Users;
