+++ src/Dfe.SignIn.Core.Contracts/Users/InitiateChangeEmailAddress.cs
@@ -1,5 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 using Dfe.SignIn.Base.Framework;
+using Dfe.SignIn.Core.Contracts.Features.Users.Exceptions;
 
 namespace Dfe.SignIn.Core.Contracts.Users;
