+++ tests/Dfe.SignIn.Core.UseCases.UnitTests/Users/ChangeJobTitleUseCaseTests.cs
@@ -1,4 +1,5 @@
 using Dfe.SignIn.Core.Contracts.Audit;
+using Dfe.SignIn.Core.Contracts.Features.Users.Exceptions;
 using Dfe.SignIn.Core.Contracts.Users;
 using Dfe.SignIn.Core.Entities.Directories;
 using Dfe.SignIn.Core.UseCases.Users;
