+++ tests/Dfe.SignIn.Core.UseCases.UnitTests/Users/GetUserOrganisationsUseCaseTests.cs
@@ -1,3 +1,4 @@
+using Dfe.SignIn.Core.Contracts.Features.Users.Exceptions;
 using Dfe.SignIn.Core.Contracts.Organisations;
 using Dfe.SignIn.Core.Contracts.Users;
 using Dfe.SignIn.Core.Public;
