+++ tests/Dfe.SignIn.Core.UseCases.UnitTests/Users/GetUserProfileUseCaseTests.cs
@@ -1,3 +1,4 @@
+using Dfe.SignIn.Core.Contracts.Features.Users.Exceptions;
 using Dfe.SignIn.Core.Contracts.Users;
 using Dfe.SignIn.Core.Entities.Directories;
 using Dfe.SignIn.Core.UseCases.Users;
